读取文件完善1
修复了0个obj时的旋转错误，同时限制了只能选取obj文件

+++ Painter.cpp
@@ -475,7 +475,7 @@ int Painter::get_mode() const {
 	return mode;  
 }
 
-int Painter::get_objnumber() const {
+size_t Painter::get_objnumber() const {
 	return obj.size();
 }
